Deep Learning & NLP Practice (Coursework)/RAG for Financial Report Analysis (LiteLLM)/requirements.txt
# Python >= 3.10 required (developed with 3.11.x)

sentence-transformers==3.0.1
chromadb==0.5.5
python-dotenv==1.0.1
pypdf==4.3.1
openai==1.75.0
litellm==1.51.0